Change tag from max-width to max-height for images
The size of images differed in height so that when they were flex-
aligned to the start, the bottom edges of the images differed with some
images being larger top-to-bottom than others. This affected the layout
and evenness of the images and of the text underneath. To correct This
I changed the tag of images from max-width to max-height. This allowed
some images to span a little wider without affecting the proportional
layout, vertically