Complete css styling for information subheading
Changed container width and margins to align everything centrally on
horizontal axis. Changed font size, family, color and boldness of subheading
and aligned it centrally. Put image and text pairs in display flex and
text-aligned centrally. Justified-content to start and aligned-items
centrally and added gaps in between flex items. Changed font-size,
family and color of text. Wrapped images in border and added roundness
to border.

This is done to look visually pleasing and make layout appealing and
to ease flow of site