Fix size of image in header using css
Before the change, the image in header was distorted to having a height
of 100px without being proportional to the width. This was as the img
type selector had max-height to 100px and width auto. However due to
separate declaration placed on the image in overview, it became
distorted. As a result, the declarations for img were all removed and
declarations were placed separately/individually on the class for image
in overview and class for the group of 4 images, respectively.