Finish styling of signpost section
For html - A sub container was required to position items within the
blue background

For css - Width for background was defined and aligned centrally on
horizontal axis using margin. Top and bottom margin of blue background
was increased for spacing as well as curved borders. Within blue
container items were displayed flex to separate button and sentences and
padding was used for position. They were aligned centrally on vertical
axis using align-items and there was space-around declaration on justify
-content. Sentences had different fonts, colors and boldness. Border was
applied to button and curved